src/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
email-validator>=2.0.0
